+++ demos/main/src/main/java/androidx/media3/demo/main/PlayerActivity.java
@@ -37,7 +37,7 @@ import androidx.media3.common.MediaItem;
 import androidx.media3.common.PlaybackException;
 import androidx.media3.common.Player;
 import androidx.media3.common.TrackSelectionParameters;
-import androidx.media3.common.TracksInfo;
+import androidx.media3.common.Tracks;
 import androidx.media3.common.util.Util;
 import androidx.media3.datasource.DataSource;
 import androidx.media3.exoplayer.ExoPlayer;
@@ -83,7 +83,7 @@ public class PlayerActivity extends AppCompatActivity
   private List<MediaItem> mediaItems;
   private TrackSelectionParameters trackSelectionParameters;
   private DebugTextViewHelper debugViewHelper;
-  private TracksInfo lastSeenTracksInfo;
+  private Tracks lastSeenTracks;
   private boolean startAutoPlay;
   private int startItemIndex;
   private long startPosition;
@@ -276,7 +276,7 @@ public class PlayerActivity extends AppCompatActivity
       RenderersFactory renderersFactory =
           DemoUtil.buildRenderersFactory(/* context= */ this, preferExtensionDecoders);
 
-      lastSeenTracksInfo = TracksInfo.EMPTY;
+      lastSeenTracks = Tracks.EMPTY;
       player =
           new ExoPlayer.Builder(/* context= */ this)
               .setRenderersFactory(renderersFactory)
@@ -456,22 +456,20 @@ public class PlayerActivity extends AppCompatActivity
 
     @Override
     @SuppressWarnings("ReferenceEquality")
-    public void onTracksInfoChanged(TracksInfo tracksInfo) {
+    public void onTracksChanged(Tracks tracks) {
       updateButtonVisibility();
-      if (tracksInfo == lastSeenTracksInfo) {
+      if (tracks == lastSeenTracks) {
         return;
       }
-      if (tracksInfo.containsType(C.TRACK_TYPE_VIDEO)
-          && !tracksInfo.isTypeSupported(
-              C.TRACK_TYPE_VIDEO, /* allowExceedsCapabilities= */ true)) {
+      if (tracks.containsType(C.TRACK_TYPE_VIDEO)
+          && !tracks.isTypeSupported(C.TRACK_TYPE_VIDEO, /* allowExceedsCapabilities= */ true)) {
         showToast(R.string.error_unsupported_video);
       }
-      if (tracksInfo.containsType(C.TRACK_TYPE_AUDIO)
-          && !tracksInfo.isTypeSupported(
-              C.TRACK_TYPE_AUDIO, /* allowExceedsCapabilities= */ true)) {
+      if (tracks.containsType(C.TRACK_TYPE_AUDIO)
+          && !tracks.isTypeSupported(C.TRACK_TYPE_AUDIO, /* allowExceedsCapabilities= */ true)) {
         showToast(R.string.error_unsupported_audio);
       }
-      lastSeenTracksInfo = tracksInfo;
+      lastSeenTracks = tracks;
     }
   }