+++ libraries/ui/src/main/java/androidx/media3/ui/PlayerControlView.java
@@ -66,8 +66,7 @@ import androidx.media3.common.Timeline;
 import androidx.media3.common.TrackGroup;
 import androidx.media3.common.TrackSelectionOverride;
 import androidx.media3.common.TrackSelectionParameters;
-import androidx.media3.common.TracksInfo;
-import androidx.media3.common.TracksInfo.TrackGroupInfo;
+import androidx.media3.common.Tracks;
 import androidx.media3.common.util.Assertions;
 import androidx.media3.common.util.RepeatModeUtil;
 import androidx.media3.common.util.UnstableApi;
@@ -1101,39 +1100,37 @@ public class PlayerControlView extends FrameLayout {
         || !player.isCommandAvailable(Player.COMMAND_SET_TRACK_SELECTION_PARAMETERS)) {
       return;
     }
-    TracksInfo tracksInfo = player.getCurrentTracksInfo();
-    audioTrackSelectionAdapter.init(
-        gatherSupportedTrackInfosOfType(tracksInfo, C.TRACK_TYPE_AUDIO));
+    Tracks tracks = player.getCurrentTracks();
+    audioTrackSelectionAdapter.init(gatherSupportedTrackInfosOfType(tracks, C.TRACK_TYPE_AUDIO));
     if (controlViewLayoutManager.getShowButton(subtitleButton)) {
-      textTrackSelectionAdapter.init(
-          gatherSupportedTrackInfosOfType(tracksInfo, C.TRACK_TYPE_TEXT));
+      textTrackSelectionAdapter.init(gatherSupportedTrackInfosOfType(tracks, C.TRACK_TYPE_TEXT));
     } else {
       textTrackSelectionAdapter.init(ImmutableList.of());
     }
   }
 
   private ImmutableList<TrackInformation> gatherSupportedTrackInfosOfType(
-      TracksInfo tracksInfo, @C.TrackType int trackType) {
-    ImmutableList.Builder<TrackInformation> tracks = new ImmutableList.Builder<>();
-    List<TrackGroupInfo> trackGroupInfos = tracksInfo.getTrackGroupInfos();
-    for (int trackGroupIndex = 0; trackGroupIndex < trackGroupInfos.size(); trackGroupIndex++) {
-      TrackGroupInfo trackGroupInfo = trackGroupInfos.get(trackGroupIndex);
-      if (trackGroupInfo.getTrackType() != trackType) {
+      Tracks tracks, @C.TrackType int trackType) {
+    ImmutableList.Builder<TrackInformation> trackInfos = new ImmutableList.Builder<>();
+    List<Tracks.Group> trackGroups = tracks.getGroups();
+    for (int trackGroupIndex = 0; trackGroupIndex < trackGroups.size(); trackGroupIndex++) {
+      Tracks.Group trackGroup = trackGroups.get(trackGroupIndex);
+      if (trackGroup.getType() != trackType) {
         continue;
       }
-      for (int trackIndex = 0; trackIndex < trackGroupInfo.length; trackIndex++) {
-        if (!trackGroupInfo.isTrackSupported(trackIndex)) {
+      for (int trackIndex = 0; trackIndex < trackGroup.length; trackIndex++) {
+        if (!trackGroup.isTrackSupported(trackIndex)) {
           continue;
         }
-        Format trackFormat = trackGroupInfo.getTrackFormat(trackIndex);
+        Format trackFormat = trackGroup.getTrackFormat(trackIndex);
         if ((trackFormat.selectionFlags & C.SELECTION_FLAG_FORCED) != 0) {
           continue;
         }
         String trackName = trackNameProvider.getTrackName(trackFormat);
-        tracks.add(new TrackInformation(tracksInfo, trackGroupIndex, trackIndex, trackName));
+        trackInfos.add(new TrackInformation(tracks, trackGroupIndex, trackIndex, trackName));
       }
     }
-    return tracks.build();
+    return trackInfos.build();
   }
 
   private void updateTimeline() {
@@ -1812,19 +1809,18 @@ public class PlayerControlView extends FrameLayout {
 
   private static final class TrackInformation {
 
-    public final TrackGroupInfo trackGroupInfo;
+    public final Tracks.Group trackGroup;
     public final int trackIndex;
     public final String trackName;
 
-    public TrackInformation(
-        TracksInfo tracksInfo, int trackGroupIndex, int trackIndex, String trackName) {
-      this.trackGroupInfo = tracksInfo.getTrackGroupInfos().get(trackGroupIndex);
+    public TrackInformation(Tracks tracks, int trackGroupIndex, int trackIndex, String trackName) {
+      this.trackGroup = tracks.getGroups().get(trackGroupIndex);
       this.trackIndex = trackIndex;
       this.trackName = trackName;
     }
 
     public boolean isSelected() {
-      return trackGroupInfo.isTrackSelected(trackIndex);
+      return trackGroup.isTrackSelected(trackIndex);
     }
   }
 
@@ -1924,7 +1920,7 @@ public class PlayerControlView extends FrameLayout {
 
     private boolean hasSelectionOverride(TrackSelectionParameters trackSelectionParameters) {
       for (int i = 0; i < tracks.size(); i++) {
-        TrackGroup trackGroup = tracks.get(i).trackGroupInfo.getTrackGroup();
+        TrackGroup trackGroup = tracks.get(i).trackGroup.getTrackGroup();
         if (trackSelectionParameters.overrides.containsKey(trackGroup)) {
           return true;
         }
@@ -1998,7 +1994,7 @@ public class PlayerControlView extends FrameLayout {
         onBindViewHolderAtZeroPosition(holder);
       } else {
         TrackInformation track = tracks.get(position - 1);
-        TrackGroup trackGroup = track.trackGroupInfo.getTrackGroup();
+        TrackGroup trackGroup = track.trackGroup.getTrackGroup();
         TrackSelectionParameters params = player.getTrackSelectionParameters();
         boolean explicitlySelected = params.overrides.get(trackGroup) != null && track.isSelected();
         holder.textView.setText(track.trackName);
@@ -2013,8 +2009,7 @@ public class PlayerControlView extends FrameLayout {
                       .setOverrideForType(
                           new TrackSelectionOverride(
                               trackGroup, ImmutableList.of(track.trackIndex)))
-                      .setTrackTypeDisabled(
-                          track.trackGroupInfo.getTrackType(), /* disabled= */ false)
+                      .setTrackTypeDisabled(track.trackGroup.getType(), /* disabled= */ false)
                       .build());
               onTrackSelection(track.trackName);
               settingsWindow.dismiss();